# backend/src/docs/openapi.yaml
openapi: 3.0.0
info:
  title: Marsad API
  description: |
    Complete REST API for Marsad — Saudi B2B Business Trust Platform

    ## Authentication
    All endpoints (except `/auth/register` and `/auth/login`) require JWT token in Authorization header:
    ```
    Authorization: Bearer <jwt_token>
    ```

    ## Error Handling
    All error responses follow this format:
    ```json
    {
      "statusCode": 400,
      "errorCode": "ERR_4002",
      "message": "Validation failed",
      "timestamp": "2024-07-18T10:30:00Z",
      "path": "/auth/login"
    }
    ```

    ## Rate Limiting
    - Rate limit: 1000 requests per hour
    - Headers: `X-RateLimit-Limit`, `X-RateLimit-Remaining`, `X-RateLimit-Reset`

  version: '1.0.0'
  contact:
    name: Marsad Support
    url: https://marsad.local
    email: support@marsad.local
  license:
    name: Proprietary
    url: https://marsad.local/license
servers:
  - url: http://localhost:3000
    description: Development
  - url: https://api.marsad.local
    description: Production
paths:
...
  /reports:
    post:
      tags:
        - Reports
      summary: Create transaction report
      description: |
        Submit a new transaction report for another company.
        Reports require review by platform moderators before approval.
      operationId: reports_create
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ReportCreateRequest'
            examples:
              success:
                summary: Valid transaction report
                value:
                  targetCompanyId: 'comp_123456'
                  dealAmountRange: '50000-100000'
                  paymentCommitment: '30'
                  delayDays: 15
                  defaulted: false
                  dealtAt: '2024-07-18T10:30:00Z'
      responses:
        '201':
          description: Report submitted successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '400':
          $ref: '#/components/responses/BadRequestError'
...
  /reports/{id}/approve:
    post:
      tags:
        - Reports
      summary: Approve report (reviewers only)
      description: Approve a pending report. Requires reviewer or admin role.
      operationId: reports_approve
      security:
        - jwt: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
          description: Report ID
      responses:
        '200':
          description: Report approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /reports/{id}/reject:
    post:
      tags:
        - Reports
      summary: Reject report (reviewers only)
      description: Reject a pending report with reason
      operationId: reports_reject
      security:
        - jwt: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
                  description: Reason for rejection
      responses:
        '200':
          description: Report rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
...
  /admin/reports:
    get:
      tags:
        - Admin
      summary: Get pending reports (admin only)
      description: Get all pending reports for admin review
      operationId: admin_getPendingReports
      security:
        - jwt: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
            default: 1
        - in: query
          name: limit
          schema:
            type: integer
            default: 20
      responses:
        '200':
          description: Pending reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportResponse'
                  total:
                    type: integer
        '401':
          $ref: '#/components/responses/UnauthorizedError'
        '403':
          $ref: '#/components/responses/ForbiddenError'
  /admin/reports/all:
    get:
      tags:
        - Admin
      summary: Get all reports (admin only)
      description: Get all reports with optional status filter
      operationId: admin_getAllReports
      security:
        - jwt: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
            enum: [pending, approved, rejected, info_requested]
      responses:
        '200':
          description: All reports
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/ReportResponse'
                  total:
                    type: integer
  /admin/reports/{id}/approve:
    patch:
      tags:
        - Admin
      summary: Approve report (admin only)
      operationId: admin_approveReport
      security:
        - jwt: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Report approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /admin/reports/{id}/reject:
    post:
      tags:
        - Admin
      summary: Reject report (admin only)
      operationId: admin_rejectReport
      security:
        - jwt: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Report rejected
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReportResponse'
  /admin/reports/batch-approve:
    post:
      tags:
        - Admin
      summary: Batch approve reports (admin only)
      operationId: admin_batchApproveReports
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reportIds:
                  type: array
                  items:
                    type: string
      responses:
        '200':
          description: Reports approved
          content:
            application/json:
              schema:
                type: object
                properties:
                  approved:
                    type: integer
                  failed:
                    type: integer
  /admin/companies:
    get:
      tags:
        - Admin
      summary: Get companies list (admin only)
      operationId: admin_getCompanies
      security:
        - jwt: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Companies list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/CompanyReport'
                  total:
                    type: integer
  /admin/companies/{id}/approve:
    post:
      tags:
        - Admin
      summary: Approve company (admin only)
      operationId: admin_approveCompany
      security:
        - jwt: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Company approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompanyReport'
  /admin/companies/{id}/reject:
    post:
      tags:
        - Admin
      summary: Reject company (admin only)
      operationId: admin_rejectCompany
      security:
        - jwt: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                reason:
                  type: string
      responses:
        '200':
          description: Company rejected
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
  /admin/users:
    get:
      tags:
        - Admin
      summary: Get users list (admin only)
      operationId: admin_getUsers
      security:
        - jwt: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: role
          schema:
            type: string
      responses:
        '200':
          description: Users list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/UserResponse'
                  total:
                    type: integer
  /admin/users/{id}/status:
    patch:
      tags:
        - Admin
      summary: Update user status (admin only)
      operationId: admin_updateUserStatus
      security:
        - jwt: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                status:
                  type: string
                  enum: [active, suspended, deleted]
      responses:
        '200':
          description: User status updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UserResponse'
  /admin/audit-logs:
    get:
      tags:
        - Admin
      summary: Get audit logs (admin only)
      operationId: admin_getAuditLogs
      security:
        - jwt: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: action
          schema:
            type: string
        - in: query
          name: entity
          schema:
            type: string
        - in: query
          name: startDate
          schema:
            type: string
            format: date-time
        - in: query
          name: endDate
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: Audit logs
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/AuditLogResponse'
                  total:
                    type: integer
  /admin/requests:
    get:
      tags:
        - Admin
      summary: Get business requests (admin only)
      operationId: admin_getBusinessRequests
      security:
        - jwt: []
      parameters:
        - in: query
          name: page
          schema:
            type: integer
        - in: query
          name: limit
          schema:
            type: integer
        - in: query
          name: status
          schema:
            type: string
      responses:
        '200':
          description: Business requests
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/BusinessRequestResponse'
  /admin/requests/{id}/approve:
    post:
      tags:
        - Admin
      summary: Approve business request (admin only)
      operationId: admin_approveBusinessRequest
      security:
        - jwt: []
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Request approved
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BusinessRequestResponse'
  /admin/bulk-upload:
    post:
      tags:
        - Admin
      summary: Bulk import companies (admin only)
      operationId: admin_bulkImportCompanies
      security:
        - jwt: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                companies:
                  type: array
                  items:
                    type: object
                    properties:
                      name:
                        type: string
                      crNumber:
                        type: string
                      sector:
                        type: string
                      city:
                        type: string
      responses:
        '200':
          description: Companies imported
          content:
            application/json:
              schema:
                type: object
                properties:
                  imported:
                    type: integer
                  failed:
                    type: integer
components:
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: |
        JWT token obtained from login endpoint.
        Include in Authorization header: `Authorization: Bearer <token>`
  responses:
    UnauthorizedError:
      description: Unauthorized - missing or invalid token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    ForbiddenError:
      description: Forbidden - insufficient permissions
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
...
    BadRequestError:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
...
    ErrorResponse:
      type: object
      required:
        - statusCode
        - errorCode
        - message
      properties:
        statusCode:
          type: integer
          example: 400
        errorCode:
          type: string
          example: 'ERR_4002'
          description: Structured error code for programmatic handling
        message:
          type: string
          example: 'Validation failed'
        details:
          type: object
          description: Additional error details
        timestamp:
          type: string
          format: date-time
        path:
          type: string

    # ============================================================================
    # AUTHENTICATION SCHEMAS
    # ============================================================================
...
    CompanyReport:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        crNumber:
          type: string
        sector:
          type: string
        city:
          type: string
        trustScore:
          type: number
          format: float
        totalReports:
          type: integer
        approvedReports:
          type: integer
        averagePaymentDelay:
          type: number
          description: Average payment delay in days
        status:
          type: string
          enum: [active, suspended, pending]
        claimedBy:
          type: string
          description: User ID if profile is claimed
        lastUpdated:
          type: string
          format: date-time
        transactionHistory:
          type: array
          items:
            type: object
            properties:
              date:
                type: string
                format: date-time
              amount:
                type: string
              paymentDelay:
                type: integer
              defaulted:
                type: boolean

    # ============================================================================
    # REPORT SCHEMAS
    # ============================================================================
    ReportCreateRequest:
      type: object
      required:
        - targetCompanyId
        - dealAmountRange
        - paymentCommitment
      properties:
        targetCompanyId:
          type: string
          description: ID of company being reported on
        dealAmountRange:
          type: string
          description: Deal amount range (e.g., '10000-50000')
        paymentCommitment:
          type: string
          description: Payment commitment period (e.g., '30')
        delayDays:
          type: integer
          description: Days payment was delayed (0 if on-time)
        defaulted:
          type: boolean
          description: Whether payment defaulted
        dealtAt:
          type: string
          format: date-time
          description: Date of transaction
    ReportResponse:
      type: object
      properties:
        id:
          type: string
        reporterCompanyId:
          type: string
        targetCompanyId:
          type: string
        dealAmountRange:
          type: string
        paymentCommitment:
          type: string
        delayDays:
          type: integer
        defaulted:
          type: boolean
        status:
          type: string
          enum: [pending, approved, rejected, info_requested]
        approvedBy:
          type: string
          description: User ID of approver
        rejectionReason:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ============================================================================
    # TRUST SCORE SCHEMAS
    # ============================================================================
...
    UserResponse:
      type: object
      properties:
        id:
          type: string
        email:
          type: string
        name:
          type: string
        role:
          type: string
          enum: [user, reviewer, platform_admin]
        status:
          type: string
          enum: [active, suspended]
        createdAt:
          type: string
          format: date-time
    AuditLogResponse:
      type: object
      properties:
        id:
          type: string
        userId:
          type: string
        action:
          type: string
        entity:
          type: string
        entityId:
          type: string
        changes:
          type: object
        timestamp:
          type: string
          format: date-time
        ipAddress:
          type: string
    BusinessRequestResponse:
      type: object
      properties:
        id:
          type: string
        type:
          type: string
        status:
          type: string
        createdAt:
          type: string
          format: date-time